Extract Combine private helpers into their own file
Per Jake's review feedback, the five Combine private helpers (shared by And,
Or, Except, Xor) move from ObservableListEx.And.cs to a dedicated
ObservableListEx.Combine.cs, matching the per-operator pattern established
for the public surface.

Audit confirms Combine is the only multi-caller private helper in
ObservableListEx partials. Byte-preserving move with no functional change.
Library builds clean on all target frameworks.
