+++ src/DynamicData/List/ObservableListEx.And.cs
@@ -94,63 +94,4 @@ public static partial class ObservableListEx
     /// </remarks>
     public static IObservable<IChangeSet<T>> And<T>(this IObservableList<ISourceList<T>> sources)
         where T : notnull => sources.Combine(CombineOperator.And);
-
-    private static IObservable<IChangeSet<T>> Combine<T>(this ICollection<IObservable<IChangeSet<T>>> sources, CombineOperator type)
-        where T : notnull
-    {
-        sources.ThrowArgumentNullExceptionIfNull(nameof(sources));
-
-        return new Combiner<T>(sources, type).Run();
-    }
-
-    private static IObservable<IChangeSet<T>> Combine<T>(this IObservable<IChangeSet<T>> source, CombineOperator type, params IObservable<IChangeSet<T>>[] others)
-        where T : notnull
-    {
-        source.ThrowArgumentNullExceptionIfNull(nameof(source));
-        others.ThrowArgumentNullExceptionIfNull(nameof(others));
-
-        if (others.Length == 0)
-        {
-            throw new ArgumentException("Must be at least one item to combine with", nameof(others));
-        }
-
-        var items = source.EnumerateOne().Union(others).ToList();
-        return new Combiner<T>(items, type).Run();
-    }
-
-    private static IObservable<IChangeSet<T>> Combine<T>(this IObservableList<ISourceList<T>> sources, CombineOperator type)
-        where T : notnull
-    {
-        sources.ThrowArgumentNullExceptionIfNull(nameof(sources));
-
-        return Observable.Create<IChangeSet<T>>(
-            observer =>
-            {
-                var changesSetList = sources.Connect().Transform(s => s.Connect()).AsObservableList();
-                var subscriber = changesSetList.Combine(type).SubscribeSafe(observer);
-                return new CompositeDisposable(changesSetList, subscriber);
-            });
-    }
-
-    private static IObservable<IChangeSet<T>> Combine<T>(this IObservableList<IObservableList<T>> sources, CombineOperator type)
-        where T : notnull
-    {
-        sources.ThrowArgumentNullExceptionIfNull(nameof(sources));
-
-        return Observable.Create<IChangeSet<T>>(
-            observer =>
-            {
-                var changesSetList = sources.Connect().Transform(s => s.Connect()).AsObservableList();
-                var subscriber = changesSetList.Combine(type).SubscribeSafe(observer);
-                return new CompositeDisposable(changesSetList, subscriber);
-            });
-    }
-
-    private static IObservable<IChangeSet<T>> Combine<T>(this IObservableList<IObservable<IChangeSet<T>>> sources, CombineOperator type)
-        where T : notnull
-    {
-        sources.ThrowArgumentNullExceptionIfNull(nameof(sources));
-
-        return new DynamicCombiner<T>(sources, type).Run();
-    }
 }
